+++ addons/debug_menu/debug_menu.gd
@@ -156,10 +156,15 @@ func update_settings_label() -> void:
 
 	var viewport := get_viewport()
 
+	# The size of the viewport rendering, which determines which resolution 3D is rendered at.
+	var viewport_render_size := Vector2i()
+
 	if viewport.content_scale_mode == Window.CONTENT_SCALE_MODE_VIEWPORT:
+		viewport_render_size = viewport.get_visible_rect().size
 		settings.text += "Viewport: %d×%d, Window: %d×%d\n" % [viewport.get_visible_rect().size.x, viewport.get_visible_rect().size.y, viewport.size.x, viewport.size.y]
 	else:
 		# Window size matches viewport size.
+		viewport_render_size = viewport.size
 		settings.text += "Viewport: %d×%d\n" % [viewport.size.x, viewport.size.y]
 
 	# Display 3D settings only if relevant.
@@ -176,8 +181,8 @@ func update_settings_label() -> void:
 		settings.text += "3D scale (%s): %d%% = %d×%d" % [
 				"Bilinear" if viewport.scaling_3d_mode == Viewport.SCALING_3D_MODE_BILINEAR else "FSR 1.0",
 				viewport.scaling_3d_scale * 100,
-				viewport.get_visible_rect().size.x * viewport.scaling_3d_scale,
-				viewport.get_visible_rect().size.y * viewport.scaling_3d_scale,
+				viewport_render_size.x * viewport.scaling_3d_scale,
+				viewport_render_size.y * viewport.scaling_3d_scale,
 		]
 
 		if not antialiasing_3d_string.is_empty():